app/requirements.txt
Django>=3.0,<4.0.2
djangorestframework==3.13.1
django-cors-headers==3.14.0
django-storages==1.13.2
djangorestframework-simplejwt==5.3.1
gunicorn==20.1.0
psycopg2==2.9.9
PyJWT==2.8.0
python-decouple==3.8
sqlparse==0.5.0
tzdata==2024.1
urllib3==1.26.16
whitenoise==6.4.0

flake8>=3.6.0,<3.7.0